Extract keyboard input action execution
Move press, key-down, and key-up workflows into a stateless executor so the keyboard use case retains only validation, setup, logging, and dispatch. Preserve input update ordering, timeout escalation, and cleanup behavior.
